refera-backend/table/views.py
from rest_framework import viewsets
from table.models import Order
from table.serializers import OrderSerializer

class OrderViewSet(viewsets.ModelViewSet):
    queryset = Order.objects.all()
    serializer_class = OrderSerializer
