refera-backend/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0     
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
Markdown==3.4.1
pytz==2022.2
sqlparse==0.4.2
tzdata==2022.1